[+] useContext across all components
